Map displays more info on restaurant
The map will now display callouts(little info windows) which contains restaurant details(image and text). Clicking on it will redirect you to the restaurant detail screen. The React webview package is required to allow android to render the image on the callout in the map screen. Also re-arranged some folders(spacer, utility, etc) so its is structured properly.